Hacker rank
A tree of PP nodes is an un-directed connected graph having P−1P−1 edges. Let us denote RR as the root node. If AA is a node such that it is at a distance of LL from RR, and BB is a node such that it is at at distance of L+1L+1 from RR and AA is connected to BB, then we call AA as the parent of BB.

Similarly, if AA is at a distance of LL from RR and BB is at a distance of L+KL+K from RR and there is a path of length KK from AA to BB, then we call AA as the KKth parent of BB.

Susan likes to play with graphs and Tree data structure is one of her favorites. She has designed a problem and wants to know if anyone can solve it. Sometimes she adds or removes a leaf node. Your task is to figure out the KKth parent of a node at any instant.